pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redenote-tools"
version = "0.1.0"
description = "基于 Spider_XHS 二次开发的小红书内容检索服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "requests>=2.32,<3.0",
  "PyExecJS>=1.5.1,<2.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.3,<9.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["service*"]

[tool.setuptools.package-data]
"service.app" = ["web/*.html", "web/*.css", "web/*.js"]
"service.app.adapters.spider_xhs" = ["assets/*.js"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
